Argv refactoring; use throw/catch for help printing
This simplifies the code significantly, also removing the need for @output. Hopefully, @long_help can be somewhat cleanly removed as well.

The drive for this is actually to allow a clean :auto_help implementation. An issue (which is not a problem right now) that there is currently in the code is that exit on error and exit on help are the same thing.